docs: redux 폴더 생성
